+++ Projects/STM32F429I-Discovery/Examples/MY_LIB/arch_driver/hal/spi_driver.h
@@ -10,9 +10,9 @@ struct spi_data {
 };
 
 struct spi_config {
-	uint32_t value_tx_timeout;
-	uint32_t value_rx_timeout;
-	uint32_t value_tx_rx_timeout;
+	uint8_t value_tx_timeout;
+	uint8_t value_rx_timeout;
+	uint8_t value_tx_rx_timeout;
 };
 
 #ifdef SPI4_CS1_DEV
